[refactor] Rename ui directory to presentation to adhere to data-domain-presentation nomenclature
