+++ app/src/main/java/dev/johnoreilly/fantasypremierleague/presentation/global/Theme.kt
@@ -1,4 +1,4 @@
-package dev.johnoreilly.fantasypremierleague.ui.global
+package dev.johnoreilly.fantasypremierleague.presentation.global
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
